Quarter maze cutout design
This is the design for the maze plywood parts that were cut out. There is extra for the walls. The cuts were made slightly bigger than necessary to make them round numbers in inches. This design was much simpler than the other design and uses less wood, but the downsides are that it is less easily expandable to a full-size maze (but still possible), and may not be as tight. There are ways to solve this that we will figure out later.